+++ docs/conf.py
@@ -22,7 +22,7 @@ copyright = '2025, Jacob Schreiber'
 author = 'Jacob Schreiber'
 
 # The full version, including alpha/beta/rc tags
-release = 'v2.0.0'
+release = 'v2.1.0'
 
 
 # -- General configuration ---------------------------------------------------